tests/test_client.py
import unittest
import numpy as np
import tensorflow as tf
from client import load_local_data, train_local_model

class TestClient(unittest.TestCase):

    def test_load_local_data(self):
        dataset = load_local_data()
        for batch in dataset.take(1):
            features, labels = batch
            self.assertIn('article_input', features)
            self.assertIn('user_input', features)
            self.assertEqual(features['article_input'].dtype, tf.int32)
            self.assertEqual(features['user_input'].dtype, tf.int32)
            self.assertEqual(labels.dtype, tf.float32)
            self.assertEqual(features['article_input'].shape[1], 1)
            self.assertEqual(features['user_input'].shape[1], 1)
            self.assertEqual(labels.shape[1], 1)

    def test_train_local_model(self):
        # Create a simple mock model
        article_input = tf.keras.Input(shape=(1,), dtype=tf.int32, name='article_input')
        user_input = tf.keras.Input(shape=(1,), dtype=tf.int32, name='user_input')
        concatenated = tf.keras.layers.Concatenate()([article_input, user_input])
        output = tf.keras.layers.Dense(1)(concatenated)
        mock_model = tf.keras.Model(inputs=[article_input, user_input], outputs=output)

        permissions = {'essential': True}
        model_updates, ranked_results = train_local_model(mock_model, permissions)

        self.assertIsInstance(model_updates, list)
        self.assertIsInstance(ranked_results, list)

if __name__ == '__main__':
    unittest.main()
